+++ client/src/components/molecules/FormInput/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as S from './style';
-import Label, { Props as LabelProps } from '../../atoms/Label';
+import Label, { Props as LabelProps } from 'components/atoms/Label';
 
 export interface Props {
   /** name of input */
