+++ client/src/components/molecules/IconBtn/index.spec.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { mount } from 'enzyme';
 
 import IconBtn from '.';
-import ExternalLinkSymbol from '../../../assets/img/external-link-black.svg';
-import ExternalLinkSymbolColored from '../../../assets/img/external-link-colored.svg';
+import ExternalLinkSymbol from 'assets/img/external-link-black.svg';
+import ExternalLinkSymbolColored from 'assets/img/external-link-colored.svg';
 const tempCircleImgSrc =
   'https://cf.festa.io/default-images/host-profiles/Profile-00047.jpg';
