+++ client/src/components/organisms/Footer/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import * as S from './style';
-import logoGray from '../../../assets/img/logo-gray.svg';
-import { FOOTER_INFO } from '../../../commons/constants/string';
-import Divider from '../../atoms/Divider';
+import logoGray from 'assets/img/logo-gray.svg';
+import { FOOTER_INFO } from 'commons/constants/string';
+import { Divider } from 'components';
 
 function Footer(): React.ReactElement {
   return (
