+++ client/src/components/organisms/SignUpForm/index.stories.tsx
@@ -9,7 +9,7 @@ import {
   SIGNUP_LAST_NAME,
   SIGNUP_PHONE_NUMBER,
   SIGNUP_PASSSWORD,
-} from '../../../commons/constants/string';
+} from 'commons/constants/string';
 
 export default {
   title: 'Organisms / SignUpForm',
