+++ client/src/pages/Login/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
 import * as S from './style';
-import IconBtn from '../../components/molecules/IconBtn';
-import { OAUTH_GOOGLE, LOGIN_SOCIAL } from '../../commons/constants/string';
-import googleSvg from '../../assets/img/google.svg';
-import LogoSvg from '../../assets/img/logo.svg';
+import { IconBtn } from 'components';
+import { OAUTH_GOOGLE, LOGIN_SOCIAL } from 'commons/constants/string';
+import googleSvg from 'assets/img/google.svg';
+import LogoSvg from 'assets/img/logo.svg';
 import LoginTemplate from './templates';
 
 const { REACT_APP_SERVER_URL } = process.env;
