+++ client/src/pages/Main/index.tsx
@@ -2,11 +2,10 @@ import React, { useState, useCallback } from 'react';
 import axios from 'axios';
 
 import MainTemplate from './template';
-import MainBanner from 'components/organisms/MainBanner';
-import CardGrid from 'components/organisms/CardGrid';
-import { Event } from '../../types/Event';
-import { useIntersect } from '../../hooks';
-import delay from '../../utils/delay';
+import { MainBanner, CardGrid } from 'components';
+import { Event } from 'types/Event';
+import { useIntersect } from 'hooks';
+import delay from 'utils/delay';
 
 function Main(): React.ReactElement {
   const [events, setEvents] = useState<Event[]>([]);
