+++ client/src/pages/Main/template/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import * as S from './style';
-import BasedTemplate from '../../BasedTemplate/templates';
+import BasedTemplate from 'pages/BasedTemplate/templates';
 
 interface Props {
   mainBanner: React.ReactNode;
