+++ client/src/pages/SignUp/store.tsx
@@ -8,13 +8,13 @@ import React, {
 import { useHistory } from 'react-router-dom';
 import axios from 'axios';
 
-import { useStateReducer } from '../../hooks/base/useStateReduter';
-import { ActionParams } from '../../types/Actions';
-import { SignUpFormState } from '../../types/States';
-import { UseStateReducer } from '../../types/CustomHooks';
-import { validatePhoneNumber, validateName } from '../../utils/validateInput';
+import { useStateReducer } from 'hooks/base/useStateReduter';
+import { ActionParams } from 'types/Actions';
+import { SignUpFormState } from 'types/States';
+import { UseStateReducer } from 'types/CustomHooks';
+import { validatePhoneNumber, validateName } from 'utils/validateInput';
 
-import { UserAccountState, UserAccountAction } from '../../stores/accountStore';
+import { UserAccountState, UserAccountAction } from 'stores/accountStore';
 
 const { REACT_APP_SERVER_URL } = process.env;
